// story-website/types/character.ts
// Copyright 2021-2026 The ESAP Project
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

export interface StoryCharacter {
  id: string;
  code: string;
  color: { primary: string; dark: string };
  order?: number;
  // 翻译字段
  name: string;
  nickname?: string;
  role: string;
  quote: string;
  description: string;
  bio?: string;
  keywords?: string[];
}